Move Planner to quasar.fs
