+++ connector/src/main/scala/quasar/qscript/qsu/ExtractFreeMap.scala
@@ -19,7 +19,7 @@ package quasar.qscript.qsu
 import slamdata.Predef.{Map => SMap, _}
 
 import quasar.RenderTreeT
-import quasar.Planner.{InternalError, PlannerErrorME}
+import quasar.fs.Planner.{InternalError, PlannerErrorME}
 import quasar.effect.NameGenerator
 import quasar.fp.symbolOrder
 import quasar.qscript.RecFreeS._
