+++ connector/src/main/scala/quasar/qscript/qsu/QAuth.scala
@@ -17,7 +17,7 @@
 package quasar.qscript.qsu
 
 import slamdata.Predef._
-import quasar.Planner.{InternalError, PlannerErrorME}
+import quasar.fs.Planner.{InternalError, PlannerErrorME}
 import quasar.ejson.implicits._
 import quasar.fp.{coproductEqual, coproductShow, symbolOrder, symbolShow}
 import quasar.qscript.FreeMap
