+++ connector/src/main/scala/quasar/qscript/qsu/ReifyAutoJoins.scala
@@ -18,7 +18,7 @@ package quasar.qscript.qsu
 
 import slamdata.Predef._
 
-import quasar.Planner, Planner.PlannerErrorME
+import quasar.fs.Planner, Planner.PlannerErrorME
 import quasar.effect.NameGenerator
 import quasar.ejson.implicits._
 import quasar.fp.coproductEqual
