+++ connector/src/main/scala/quasar/qscript/qsu/ReifyBuckets.scala
@@ -19,7 +19,7 @@ package quasar.qscript.qsu
 import slamdata.Predef._
 
 import quasar.effect.NameGenerator
-import quasar.Planner.PlannerErrorME
+import quasar.fs.Planner.PlannerErrorME
 import quasar.contrib.scalaz.MonadState_
 import quasar.fp.symbolOrder
 import quasar.fp.ski.ι
