+++ connector/src/main/scala/quasar/qscript/rewrites/ExpandDirs.scala
@@ -17,7 +17,7 @@
 package quasar.qscript.rewrites
 
 import slamdata.Predef.{Map => _, _}
-import quasar.Planner.NoFilesFound
+import quasar.fs.Planner.NoFilesFound
 import quasar.contrib.pathy._
 import quasar.contrib.scalaz._
 import quasar.fp._
