+++ couchbase/src/main/scala/quasar/physical/couchbase/planner/EquiJoinPlanner.scala
@@ -18,7 +18,7 @@ package quasar.physical.couchbase.planner
 
 import slamdata.Predef._
 import quasar.effect.NameGenerator
-import quasar.Planner.{PlannerErrorME}
+import quasar.fs.Planner.{PlannerErrorME}
 import quasar.common.JoinType
 import quasar.contrib.pathy.AFile
 import quasar.fp.ski.κ
