+++ couchbase/src/main/scala/quasar/physical/couchbase/planner/Planner.scala
@@ -19,7 +19,7 @@ package quasar.physical.couchbase.planner
 import quasar.effect.NameGenerator
 import quasar.contrib.pathy.{ADir, AFile}
 import quasar.physical.couchbase._, common._
-import quasar.Planner.PlannerErrorME
+import quasar.fs.Planner.PlannerErrorME
 import quasar.qscript._
 
 import matryoshka._
