+++ couchbase/src/main/scala/quasar/physical/couchbase/planner/QScriptCorePlanner.scala
@@ -18,7 +18,7 @@ package quasar.physical.couchbase.planner
 
 import slamdata.Predef._
 import quasar.{Data => QData}
-import quasar.Planner.{PlannerErrorME, InternalError}
+import quasar.fs.Planner.{PlannerErrorME, InternalError}
 import quasar.effect.NameGenerator
 import quasar.ejson
 import quasar.fp._
