+++ couchbase/src/main/scala/quasar/physical/couchbase/planner/package.scala
@@ -18,8 +18,8 @@ package quasar.physical.couchbase
 
 import slamdata.Predef._
 import quasar.effect.NameGenerator
-import quasar.Planner.InternalError
-import quasar.Planner.PlannerErrorME
+import quasar.fs.Planner.InternalError
+import quasar.fs.Planner.PlannerErrorME
 
 import matryoshka._
 import matryoshka.implicits._
