+++ couchbase/src/test/scala/quasar/physical/couchbase/BasicQueryEnablementSpec.scala
@@ -17,14 +17,14 @@
 package quasar.physical.couchbase
 
 import slamdata.Predef._
-import quasar.{Planner => _, _}
+import quasar._
 import quasar.contrib.pathy._
 import quasar.contrib.scalaz.eitherT._
 import quasar.effect.MonotonicSeq
 import quasar.fp._
 import quasar.fp.ski.ι
 import quasar.frontend.logicalplan.LogicalPlan
-import quasar.Planner.PlannerError
+import quasar.fs.Planner.PlannerError
 import quasar.qscript.{Map => _, Read => _, _}
 import quasar.sql._
