+++ mongodb/src/main/scala/quasar/physical/mongodb/workflowbuilder.scala
@@ -17,12 +17,13 @@
 package quasar.physical.mongodb
 
 import slamdata.Predef._
-import quasar.{NonTerminal, RenderTree, Terminal}, RenderTree.ops._
+import quasar._
+import quasar.RenderTree.ops._
 import quasar.common.SortDir
 import quasar.contrib.matryoshka._
 import quasar.contrib.scalaz._
 import quasar.fp._
-import quasar._, Planner._
+import quasar.fs.Planner, Planner._
 import quasar.javascript._
 import quasar.jscore, jscore.JsFn
 import quasar.physical.mongodb.accumulator._
